+++ packages/frameworks/angular/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @qualcomm-ui/angular
 
+## 1.6.0 (2025/11/30)
+
+### Features
+
+- wire highlightMatchingText in angular combobox
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/angular-core]
+
 ## 1.5.0 (2025/11/25)
 
 ### Features